pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "excel_pg_importer"
version = "0.1.0"
description = "Excel -> PostgreSQL bulk import CLI tool"
authors = [{name="IWK", email="dev@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "psycopg2-binary>=2.9.9",
  "PyYAML>=6.0.1",
  "tqdm>=4.66.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyYAML",
  "types-psycopg2",
  "jsonschema>=4.22.0"
]

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=90"
python_files = "test_*.py"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "A"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
show_error_codes = true

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true
